+++ hutool-http/src/main/java/cn/hutool/http/server/action/RootAction.java
@@ -64,6 +64,7 @@ public class RootAction implements Action {
 	@Override
 	public void doAction(HttpServerRequest request, HttpServerResponse response) {
 		final String path = request.getPath();
+
 		File file = FileUtil.file(rootDir, path);
 		if (file.exists()) {
 			if (file.isDirectory()) {
@@ -75,7 +76,8 @@ public class RootAction implements Action {
 					}
 				}
 			} else{
-				response.write(file);
+				final String name = request.getParam("name");
+				response.write(file, name);
 			}
 		}